requirements.txt
Django
azure-storage-blob 
mindee
pandas
openpyxl